+++ seed/python-sdk/audiences/src/seed/client.py
@@ -10,6 +10,16 @@ from .resources.foo.client import AsyncFooClient, FooClient
 
 
 class SeedAudiences:
+    """
+    Use this class to access the different functions within the SDK. You can instantiate any number of clients with different configuration that will propogate to these functions.
+    ---
+    from seed.client import SeedAudiences
+
+    client = SeedAudiences(
+        base_url="https://yourhost.com/path/to/api",
+    )
+    """
+
     def __init__(
         self, *, base_url: str, timeout: typing.Optional[float] = 60, httpx_client: typing.Optional[httpx.Client] = None
     ):
@@ -21,6 +31,16 @@ class SeedAudiences:
 
 
 class AsyncSeedAudiences:
+    """
+    Use this class to access the different functions within the SDK. You can instantiate any number of clients with different configuration that will propogate to these functions.
+    ---
+    from seed.client import AsyncSeedAudiences
+
+    client = AsyncSeedAudiences(
+        base_url="https://yourhost.com/path/to/api",
+    )
+    """
+
     def __init__(
         self,
         *,