+++ seed/python-sdk/auth-environment-variables/src/seed/client.py
@@ -11,6 +11,17 @@ from .resources.service.client import AsyncServiceClient, ServiceClient
 
 
 class SeedAuthEnvironmentVariables:
+    """
+    Use this class to access the different functions within the SDK. You can instantiate any number of clients with different configuration that will propogate to these functions.
+    ---
+    from seed.client import SeedAuthEnvironmentVariables
+
+    client = SeedAuthEnvironmentVariables(
+        api_key="YOUR_API_KEY",
+        base_url="https://yourhost.com/path/to/api",
+    )
+    """
+
     def __init__(
         self,
         *,
@@ -32,6 +43,17 @@ class SeedAuthEnvironmentVariables:
 
 
 class AsyncSeedAuthEnvironmentVariables:
+    """
+    Use this class to access the different functions within the SDK. You can instantiate any number of clients with different configuration that will propogate to these functions.
+    ---
+    from seed.client import AsyncSeedAuthEnvironmentVariables
+
+    client = AsyncSeedAuthEnvironmentVariables(
+        api_key="YOUR_API_KEY",
+        base_url="https://yourhost.com/path/to/api",
+    )
+    """
+
     def __init__(
         self,
         *,