+++ seed/python-sdk/enum/src/seed/client.py
@@ -11,6 +11,16 @@ from .resources.query_param.client import AsyncQueryParamClient, QueryParamClien
 
 
 class SeedEnum:
+    """
+    Use this class to access the different functions within the SDK. You can instantiate any number of clients with different configuration that will propogate to these functions.
+    ---
+    from seed.client import SeedEnum
+
+    client = SeedEnum(
+        base_url="https://yourhost.com/path/to/api",
+    )
+    """
+
     def __init__(
         self, *, base_url: str, timeout: typing.Optional[float] = 60, httpx_client: typing.Optional[httpx.Client] = None
     ):
@@ -23,6 +33,16 @@ class SeedEnum:
 
 
 class AsyncSeedEnum:
+    """
+    Use this class to access the different functions within the SDK. You can instantiate any number of clients with different configuration that will propogate to these functions.
+    ---
+    from seed.client import AsyncSeedEnum
+
+    client = AsyncSeedEnum(
+        base_url="https://yourhost.com/path/to/api",
+    )
+    """
+
     def __init__(
         self,
         *,