+++ seed/python-sdk/error-property/src/seed/client.py
@@ -9,6 +9,16 @@ from .resources.property_based_error.client import AsyncPropertyBasedErrorClient
 
 
 class SeedErrorProperty:
+    """
+    Use this class to access the different functions within the SDK. You can instantiate any number of clients with different configuration that will propogate to these functions.
+    ---
+    from seed.client import SeedErrorProperty
+
+    client = SeedErrorProperty(
+        base_url="https://yourhost.com/path/to/api",
+    )
+    """
+
     def __init__(
         self, *, base_url: str, timeout: typing.Optional[float] = 60, httpx_client: typing.Optional[httpx.Client] = None
     ):
@@ -19,6 +29,16 @@ class SeedErrorProperty:
 
 
 class AsyncSeedErrorProperty:
+    """
+    Use this class to access the different functions within the SDK. You can instantiate any number of clients with different configuration that will propogate to these functions.
+    ---
+    from seed.client import AsyncSeedErrorProperty
+
+    client = AsyncSeedErrorProperty(
+        base_url="https://yourhost.com/path/to/api",
+    )
+    """
+
     def __init__(
         self,
         *,