+++ seed/python-sdk/examples/src/seed/client.py
@@ -25,6 +25,18 @@ OMIT = typing.cast(typing.Any, ...)
 
 
 class SeedExamples:
+    """
+    Use this class to access the different functions within the SDK. You can instantiate any number of clients with different configuration that will propogate to these functions.
+    ---
+    from seed.client import SeedExamples
+    from seed.environment import SeedExamplesEnvironment
+
+    client = SeedExamples(
+        token="YOUR_TOKEN",
+        environment=SeedExamplesEnvironment.PRODUCTION,
+    )
+    """
+
     def __init__(
         self,
         *,
@@ -90,6 +102,18 @@ class SeedExamples:
 
 
 class AsyncSeedExamples:
+    """
+    Use this class to access the different functions within the SDK. You can instantiate any number of clients with different configuration that will propogate to these functions.
+    ---
+    from seed.client import AsyncSeedExamples
+    from seed.environment import SeedExamplesEnvironment
+
+    client = AsyncSeedExamples(
+        token="YOUR_TOKEN",
+        environment=SeedExamplesEnvironment.PRODUCTION,
+    )
+    """
+
     def __init__(
         self,
         *,