+++ seed/python-sdk/folders/src/seed/client.py
@@ -15,6 +15,16 @@ from .resources.folder.client import AsyncFolderClient, FolderClient
 
 
 class SeedApi:
+    """
+    Use this class to access the different functions within the SDK. You can instantiate any number of clients with different configuration that will propogate to these functions.
+    ---
+    from seed.client import SeedApi
+
+    client = SeedApi(
+        base_url="https://yourhost.com/path/to/api",
+    )
+    """
+
     def __init__(
         self, *, base_url: str, timeout: typing.Optional[float] = 60, httpx_client: typing.Optional[httpx.Client] = None
     ):
@@ -60,6 +70,16 @@ class SeedApi:
 
 
 class AsyncSeedApi:
+    """
+    Use this class to access the different functions within the SDK. You can instantiate any number of clients with different configuration that will propogate to these functions.
+    ---
+    from seed.client import AsyncSeedApi
+
+    client = AsyncSeedApi(
+        base_url="https://yourhost.com/path/to/api",
+    )
+    """
+
     def __init__(
         self,
         *,