+++ seed/python-sdk/literal/src/seed/client.py
@@ -13,6 +13,16 @@ from .resources.reference.client import AsyncReferenceClient, ReferenceClient
 
 
 class SeedLiteral:
+    """
+    Use this class to access the different functions within the SDK. You can instantiate any number of clients with different configuration that will propogate to these functions.
+    ---
+    from seed.client import SeedLiteral
+
+    client = SeedLiteral(
+        base_url="https://yourhost.com/path/to/api",
+    )
+    """
+
     def __init__(
         self, *, base_url: str, timeout: typing.Optional[float] = 60, httpx_client: typing.Optional[httpx.Client] = None
     ):
@@ -27,6 +37,16 @@ class SeedLiteral:
 
 
 class AsyncSeedLiteral:
+    """
+    Use this class to access the different functions within the SDK. You can instantiate any number of clients with different configuration that will propogate to these functions.
+    ---
+    from seed.client import AsyncSeedLiteral
+
+    client = AsyncSeedLiteral(
+        base_url="https://yourhost.com/path/to/api",
+    )
+    """
+
     def __init__(
         self,
         *,