+++ seed/python-sdk/object/src/seed/client.py
@@ -8,6 +8,16 @@ from .core.client_wrapper import AsyncClientWrapper, SyncClientWrapper
 
 
 class SeedObject:
+    """
+    Use this class to access the different functions within the SDK. You can instantiate any number of clients with different configuration that will propogate to these functions.
+    ---
+    from seed.client import SeedObject
+
+    client = SeedObject(
+        base_url="https://yourhost.com/path/to/api",
+    )
+    """
+
     def __init__(
         self, *, base_url: str, timeout: typing.Optional[float] = 60, httpx_client: typing.Optional[httpx.Client] = None
     ):
@@ -17,6 +27,16 @@ class SeedObject:
 
 
 class AsyncSeedObject:
+    """
+    Use this class to access the different functions within the SDK. You can instantiate any number of clients with different configuration that will propogate to these functions.
+    ---
+    from seed.client import AsyncSeedObject
+
+    client = AsyncSeedObject(
+        base_url="https://yourhost.com/path/to/api",
+    )
+    """
+
     def __init__(
         self,
         *,