+++ seed/python-sdk/package-yml/src/seed/client.py
@@ -23,6 +23,16 @@ OMIT = typing.cast(typing.Any, ...)
 
 
 class SeedPackageYml:
+    """
+    Use this class to access the different functions within the SDK. You can instantiate any number of clients with different configuration that will propogate to these functions.
+    ---
+    from seed.client import SeedPackageYml
+
+    client = SeedPackageYml(
+        base_url="https://yourhost.com/path/to/api",
+    )
+    """
+
     def __init__(
         self, *, base_url: str, timeout: typing.Optional[float] = 60, httpx_client: typing.Optional[httpx.Client] = None
     ):
@@ -79,6 +89,16 @@ class SeedPackageYml:
 
 
 class AsyncSeedPackageYml:
+    """
+    Use this class to access the different functions within the SDK. You can instantiate any number of clients with different configuration that will propogate to these functions.
+    ---
+    from seed.client import AsyncSeedPackageYml
+
+    client = AsyncSeedPackageYml(
+        base_url="https://yourhost.com/path/to/api",
+    )
+    """
+
     def __init__(
         self,
         *,