+++ seed/python-sdk/plain-text/src/seed/client.py
@@ -9,6 +9,16 @@ from .resources.service.client import AsyncServiceClient, ServiceClient
 
 
 class SeedPlainText:
+    """
+    Use this class to access the different functions within the SDK. You can instantiate any number of clients with different configuration that will propogate to these functions.
+    ---
+    from seed.client import SeedPlainText
+
+    client = SeedPlainText(
+        base_url="https://yourhost.com/path/to/api",
+    )
+    """
+
     def __init__(
         self, *, base_url: str, timeout: typing.Optional[float] = 60, httpx_client: typing.Optional[httpx.Client] = None
     ):
@@ -19,6 +29,16 @@ class SeedPlainText:
 
 
 class AsyncSeedPlainText:
+    """
+    Use this class to access the different functions within the SDK. You can instantiate any number of clients with different configuration that will propogate to these functions.
+    ---
+    from seed.client import AsyncSeedPlainText
+
+    client = AsyncSeedPlainText(
+        base_url="https://yourhost.com/path/to/api",
+    )
+    """
+
     def __init__(
         self,
         *,